+++ libNbiot/src/nbiotmqttsnclient.cpp
@@ -115,7 +115,7 @@ bool NbiotMqttSnClient::loopWait(int breakValue, int interval)
     }
     else
     {
-        nbiot::Timer timer(static_cast<unsigned long>(interval));
+        nbiot::Timer timer(interval);
 
         if(breakValue == cycle(timer))
         {
@@ -156,7 +156,7 @@ bool NbiotMqttSnClient::startYieldLoop(int& loopTime)
 #endif
     debugPrintf("MQTT::yield start %d\r\n", loopTime);
 #endif
-    m_yieldLoopClient.getTimer().start(static_cast<unsigned long>(loopTime));
+    m_yieldLoopClient.getTimer().start(loopTime);
     yieldRunning = true;
     return true;
 }
@@ -218,7 +218,7 @@ bool NbiotMqttSnClient::startConLoop(int& len)
     if (MQTTSN::SUCCESS == sendPacket(len, timer)) // send the connect packet
     {
         if (this->duration > 0)
-            last_received.countdown(static_cast<unsigned long>(this->duration));
+            last_received.countdown(this->duration);
 
         m_conLoopClient.getTimer().start(timer.remaining());
         ret = true;
@@ -326,7 +326,7 @@ bool NbiotMqttSnClient::finishConLoop(int& loopTime)
     if (cleansession==0 && inflightMsgid>0)
     {
         memcpy(sendbuf, pubbuf, getMaxPacketSize());
-        nbiot::Timer connect_timer = nbiot::Timer(static_cast<unsigned long>(loopTime));
+        nbiot::Timer connect_timer = nbiot::Timer(loopTime);
         if(MQTTSN::SUCCESS != publish(inflightLen, connect_timer, inflightQoS))
             ret = false;
     }
@@ -617,7 +617,7 @@ int NbiotMqttSnClient::pmPing(char *cID)
     {
         MQTTSNString clientid = MQTTSNString_initializer;
         clientid.cstring = cID;
-        nbiot::Timer timer = nbiot::Timer((unsigned) 1000);
+        nbiot::Timer timer = nbiot::Timer(1000);
         int len = MQTTSNSerialize_pingreq(sendbuf, getMaxPacketSize(), clientid);
         if (len > 0 && (rc = sendPacket(len, timer)) == MQTTSN::SUCCESS) // send the ping packet
             pmPingResp_outstanding = true;
